backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-dotenv==1.0.1
google-generativeai==0.8.3
google-cloud-aiplatform==1.75.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
python-multipart==0.0.17
pydantic==2.10.3
pydantic-settings==2.6.1
aiofiles==24.1.0
pillow==11.0.0
httpx==0.27.2
